Extract auth testing utils from 'test_auth_methods-t.cpp'
